remove immutability constraints on dataclasses
