--- gw_spaceheat/data_classes/cacs/electric_heater_cac_base.py
@@ -1,45 +1,0 @@
-"""ElectricHeaterCacBase definition"""
-
-from abc import abstractmethod
-from typing import Optional, Dict
-
-from schema.gt.gt_electric_heater_cac.gt_electric_heater_cac import GtElectricHeaterCac
-from data_classes.component_attribute_class import ComponentAttributeClass
-from data_classes.errors import DcError
-from schema.enums.make_model.make_model_map import MakeModelMap
-
-
-class ElectricHeaterCacBase(ComponentAttributeClass):
-    base_props = []
-    
-    base_props.append("component_attribute_class_id")
-    base_props.append("make_model")
-    base_props.append("display_name")
-
-    def __init__(self, component_attribute_class_id: str,
-                 make_model_gt_enum_symbol: str,
-                 display_name: Optional[str] = None,
-                 ):
-
-        super(ElectricHeaterCacBase, self).__init__(component_attribute_class_id=component_attribute_class_id,
-                                             display_name=display_name)
-        self.make_model = MakeModelMap.gt_to_local(make_model_gt_enum_symbol)
-
-    def update(self, type: GtElectricHeaterCac):
-        self._check_immutability_constraints(type=type)
-
-    def _check_immutability_constraints(self, type: GtElectricHeaterCac):
-        if self.component_attribute_class_id != type.ComponentAttributeClassId:
-            raise DcError(f'component_attribute_class_id must be immutable for {self}. '
-                          f'Got {type.ComponentAttributeClassId}')
-        if self.make_model != type.MakeModel:
-            raise DcError(f'make_model must be immutable for {self}. '
-                          f'Got {type.MakeModel}')
-
-    @abstractmethod
-    def _check_update_axioms(self, type: GtElectricHeaterCac):
-        raise NotImplementedError
-
-    @abstractmethod
-    def __repr__(self):
-        raise NotImplementedError
