+++ gw_spaceheat/data_classes/components/electric_heater_component.py
@@ -1,29 +1,37 @@
 """ElectricHeaterComponent definition"""
 from typing import Dict, Optional
 
+from data_classes.cacs.electric_heater_cac import ElectricHeaterCac
 from data_classes.component import Component
-from data_classes.components.electric_heater_component_base import ElectricHeaterComponentBase
-from schema.gt.gt_electric_heater_component.gt_electric_heater_component import GtElectricHeaterComponent
+from schema.enums.make_model.make_model_map import MakeModel
 
 
-class ElectricHeaterComponent(ElectricHeaterComponentBase):
-    by_id: Dict[str, "ElectricHeaterComponent"] =  {}
+class ElectricHeaterComponent(Component):
+    by_id: Dict[str, "ElectricHeaterComponent"] = {}
 
-    def __init__(self, component_id: str,
-                 component_attribute_class_id: str,
-                 hw_uid: Optional[str] = None,
-                 display_name: Optional[str] = None,
-                 ):
-        super(self.__class__, self).__init__(hw_uid=hw_uid,
-                                             display_name=display_name,
-                                             component_id=component_id,
-                                             component_attribute_class_id=component_attribute_class_id,
-                                             )
+    def __init__(
+        self,
+        component_id: str,
+        component_attribute_class_id: str,
+        hw_uid: Optional[str] = None,
+        display_name: Optional[str] = None,
+    ):
+        super(self.__class__, self).__init__(
+            display_name=display_name,
+            component_id=component_id,
+            hw_uid=hw_uid,
+            component_attribute_class_id=component_attribute_class_id,
+        )
         ElectricHeaterComponent.by_id[self.component_id] = self
         Component.by_id[self.component_id] = self
 
-    def _check_update_axioms(self, type: GtElectricHeaterComponent):
-        pass
+    @property
+    def cac(self) -> ElectricHeaterCac:
+        return ElectricHeaterCac.by_id[self.component_attribute_class_id]
+
+    @property
+    def make_model(self) -> MakeModel:
+        return self.cac.make_model
 
     def __repr__(self):
         return f"{self.display_name}  ({self.cac.make_model.value})"
