+++ gw_spaceheat/data_classes/components/electric_meter_component.py
@@ -1,14 +1,12 @@
 """ElectricMeterComponent definition"""
 from typing import Dict, Optional
 
+from data_classes.cacs.electric_meter_cac import ElectricMeterCac
 from data_classes.component import Component
-from data_classes.components.electric_meter_component_base import ElectricMeterComponentBase
-from schema.gt.gt_electric_meter_component.gt_electric_meter_component import (
-    GtElectricMeterComponent,
-)
+from schema.enums.make_model.make_model_map import MakeModel
 
 
-class ElectricMeterComponent(ElectricMeterComponentBase):
+class ElectricMeterComponent(Component):
     by_id: Dict[str, "ElectricMeterComponent"] = {}
 
     def __init__(
@@ -24,11 +22,17 @@ class ElectricMeterComponent(ElectricMeterComponentBase):
             hw_uid=hw_uid,
             component_attribute_class_id=component_attribute_class_id,
         )
+
         ElectricMeterComponent.by_id[self.component_id] = self
         Component.by_id[self.component_id] = self
 
-    def _check_update_axioms(self, gw_tuple: GtElectricMeterComponent):
-        pass
+    @property
+    def cac(self) -> ElectricMeterCac:
+        return ElectricMeterCac.by_id[self.component_attribute_class_id]
+
+    @property
+    def make_model(self) -> MakeModel:
+        return self.cac.make_model
 
     def __repr__(self):
         return f"{self.display_name}  ({self.cac.make_model.value})"
