+++ gw_spaceheat/data_classes/components/temp_sensor_component.py
@@ -1,29 +1,37 @@
 """TempSensorComponent definition"""
 from typing import Dict, Optional
 
+from data_classes.cacs.temp_sensor_cac import TempSensorCac
 from data_classes.component import Component
-from data_classes.components.temp_sensor_component_base import TempSensorComponentBase
-from schema.gt.gt_temp_sensor_component.gt_temp_sensor_component import GtTempSensorComponent
+from schema.enums.make_model.make_model_map import MakeModel
 
 
-class TempSensorComponent(TempSensorComponentBase):
-    by_id: Dict[str, "TempSensorComponent"] =  {}
+class TempSensorComponent(Component):
+    by_id: Dict[str, "TempSensorComponent"] = {}
 
-    def __init__(self, component_id: str,
-                 component_attribute_class_id: str,
-                 display_name: Optional[str] = None,
-                 hw_uid: Optional[str] = None,
-                 ):
-        super(self.__class__, self).__init__(display_name=display_name,
-                                             component_id=component_id,
-                                             hw_uid=hw_uid,
-                                             component_attribute_class_id=component_attribute_class_id,
-                                             )
+    def __init__(
+        self,
+        component_id: str,
+        component_attribute_class_id: str,
+        display_name: Optional[str] = None,
+        hw_uid: Optional[str] = None,
+    ):
+        super(self.__class__, self).__init__(
+            display_name=display_name,
+            component_id=component_id,
+            hw_uid=hw_uid,
+            component_attribute_class_id=component_attribute_class_id,
+        )
         TempSensorComponent.by_id[self.component_id] = self
         Component.by_id[self.component_id] = self
 
-    def _check_update_axioms(self, type: GtTempSensorComponent):
-        pass
+    @property
+    def cac(self) -> TempSensorCac:
+        return TempSensorCac.by_id[self.component_attribute_class_id]
+
+    @property
+    def make_model(self) -> MakeModel:
+        return self.cac.make_model
 
     def __repr__(self):
         return f"{self.display_name}  ({self.cac.make_model.value})"
