--- gw_spaceheat/data_classes/sh_node_base.py
@@ -1,53 +1,0 @@
-"""ShNodeBase definition"""
-
-from abc import ABC, abstractmethod
-from typing import Optional
-
-from schema.gt.gt_sh_node.gt_sh_node import GtShNode
-from data_classes.errors import DcError
-from schema.enums.role.role_map import RoleMap
-from schema.enums.actor_class.actor_class_map import ActorClassMap
-
-
-class ShNodeBase(ABC):
-    base_props = []
-    base_props.append("sh_node_id")
-    base_props.append("alias")
-    base_props.append("component_id")
-    base_props.append("reporting_sample_period_s")
-    base_props.append("display_name")
-    base_props.append("actor_class_gt_enum_symbol")
-    base_props.append("role_gt_enum_symbol")
-
-    def __init__(self,
-                 sh_node_id: str,
-                 alias: str,
-                 role_gt_enum_symbol: str,
-                 actor_class_gt_enum_symbol: str,
-                 reporting_sample_period_s: Optional[int] = None,
-                 component_id: Optional[str] = None,
-                 display_name: Optional[str] = None,
-                 ):
-        self.sh_node_id = sh_node_id
-        self.alias = alias
-        self.component_id = component_id
-        self.reporting_sample_period_s = reporting_sample_period_s
-        self.display_name = display_name
-        self.actor_class = ActorClassMap.gt_to_local(actor_class_gt_enum_symbol)
-        self.role = RoleMap.gt_to_local(role_gt_enum_symbol)
-
-    def update(self, type: GtShNode):
-        self._check_immutability_constraints(type=type)
-
-    def _check_immutability_constraints(self, type: GtShNode):
-        if self.role != type.Role:
-            raise DcError(f'role must be immutable for {self}. '
-                          f'Got {type.Role}')
-
-    @abstractmethod
-    def _check_update_axioms(self, type: GtShNode):
-        raise NotImplementedError
-
-    @abstractmethod
-    def __repr__(self):
-        raise NotImplementedError
